Fix test-soft-shutdown-cluster
