+++ tests/js/client/restart/test-soft-shutdown-cluster.js
@@ -28,63 +28,19 @@
 /// @author Copyright 2021, ArangoDB Inc, Cologne, Germany
 //////////////////////////////////////////////////////////////////////////////
 
-const _ = require('lodash');
 let jsunity = require('jsunity');
-const pu = require('@arangodb/testutils/process-utils');
-const request = require("@arangodb/request");
 const internal = require("internal");
 const db = require("internal").db;
 const time = internal.time;
 const wait = internal.wait;
-const statusExternal = internal.statusExternal;
 const {
   getCtrlCoordinators
 } = require('@arangodb/test-helper');
 
-var graph_module = require("@arangodb/general-graph");
-var EPS = 0.0001;
-let pregel = require("@arangodb/pregel");
-
-const graphName = "UnitTest_pregel";
-const vColl = "UnitTest_pregel_v", eColl = "UnitTest_pregel_e";
-
-function testAlgoStart(a, p) {
-  let pid = pregel.start(a, graphName, p);
-  assertTrue(typeof pid === "string");
-  return pid;
-}
-
-function testAlgoCheck(pid) {
-  let stats = pregel.status(pid);
-  console.warn("Pregel status:", stats);
-  return stats.state !== "running" && stats.state !== "storing";
-}
-
-function waitForAlive(timeout, baseurl, data) {
-  let res;
-  let all = Object.assign(data || {}, { method: "get", timeout: 1, url: baseurl + "/_api/version" });
-  const end = time() + timeout;
-  while (time() < end) {
-    res = request(all);
-    if (res.status === 200 || res.status === 401 || res.status === 403) {
-      break;
-    }
-    console.warn("waiting for server response from url " + baseurl);
-    wait(0.5);
-  }
-  return res.status;
-};
-
-function restartInstance(arangod) {
-  let options = _.clone(global.obj.options);
-  options.skipReconnect = false;
-  arangod.restartOneInstance();
-  waitForAlive(30, arangod.url, {});
-  arangod.checkArangoConnection(5);
-};
-
 function testSuite() {
   let cn = "UnitTestSoftShutdown";
+
+  let coordinator;
 
   return {
     setUp : function() {
@@ -95,35 +51,38 @@ function testSuite() {
         docs.push({Hallo:i});
       }
       collection.save(docs);
+
+      let coordinators = getCtrlCoordinators();
+      assertTrue(coordinators.length > 0, "expect coordinators.length > 0");
+
+      // TODO: Pick a random coordinator?
+      coordinator = coordinators[0];
     },
 
     tearDown : function() {
+      // Every test is supposed to initiate a soft shutdown
+      // hence, once we tear down the test we expect
+      // the coordinator to be gone
+      coordinator.waitForInstanceShutdown(30);
+      coordinator.restartOneInstance();
+      coordinator.checkArangoConnection(10);
+      // we might have been connected to the coordinator that
+      // has been shut down
       db._drop(cn);
     },
 
     testSoftShutdownWithoutTraffic : function() {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
       // Now use soft shutdown API to shut coordinator down:
       let status = arango.GET("/_admin/shutdown");
-      assertFalse(status.softShutdownOngoing);
+      assertFalse(status.softShutdownOngoing, "expect status.softShutdownOngoing == false");
       let res = arango.DELETE("/_admin/shutdown?soft=true");
-      assertEqual("OK", res);
+      assertEqual("OK", res, `expect res = ${res} == \"OK\"`);
       status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-      assertEqual(0, status.AQLcursors);
-      assertEqual(0, status.transactions);
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
+      assertTrue(status.softShutdownOngoing, `expect status.softShutdownOngoing = ${status.softShutdownOngoing} == true`);
+      assertEqual(0, status.AQLcursors, "expect status.AQLcursors == 0");
     },
 
     testSoftShutdownWithAQLCursor : function() {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
       // Create an AQL cursor:
       let data = {
         query: `FOR x in ${cn} RETURN x`,
@@ -137,8 +96,9 @@ function testSuite() {
       assertEqual("OK", res);
       let status = arango.GET("/_admin/shutdown");
       assertTrue(status.softShutdownOngoing);
-      assertEqual(1, status.AQLcursors);
-      assertEqual(1, status.transactions);
+      // Expect our cursor and transaction to still be there
+      assertTrue(status.AQLcursors >= 1, `expect status.AQLcursors = ${status.AQLcursors} >= 1`);
+      assertTrue(status.transactions >= 1, `expect status.transactions = ${status.transactions} >= 1`);
 
       // It should fail to create a new cursor:
       let respFailed = arango.POST("/_api/cursor", data);
@@ -159,17 +119,9 @@ function testSuite() {
       assertFalse(next.hasMore);
       assertFalse(next.error);
       assertEqual(200, next.code);
-
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
     },
 
     testSoftShutdownWithAQLCursorDeleted : function() {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
       // Create an AQL cursor:
       let data = {
         query: `FOR x in ${cn} RETURN x`,
@@ -186,8 +138,8 @@ function testSuite() {
       assertEqual("OK", res);
       status = arango.GET("/_admin/shutdown");
       assertTrue(status.softShutdownOngoing);
-      assertEqual(1, status.AQLcursors);
-      assertEqual(1, status.transactions);
+      assertTrue(status.AQLcursors >= 1, `expect status.AQLcursors = ${status.AQLcursors} >= 1`);
+      assertTrue(status.transactions >= 1, `expect status.transactions = ${status.transactions} >= 1`);
 
       // It should fail to create a new cursor:
       let respFailed = arango.POST("/_api/cursor", data);
@@ -208,17 +160,9 @@ function testSuite() {
       assertFalse(next.hasMore);
       assertFalse(next.error);
       assertEqual(202, next.code);
-
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
     },
 
     testSoftShutdownWithStreamingTrx : function() {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
       // Create a streaming transaction:
       let data = { collections: {write: [cn]} };
 
@@ -231,9 +175,8 @@ function testSuite() {
       let res = arango.DELETE("/_admin/shutdown?soft=true");
       assertEqual("OK", res);
       status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-      assertEqual(0, status.AQLcursors);
-      assertEqual(1, status.transactions);
+      assertTrue(status.softShutdownOngoing, "expect status.softShutdownOngoing == true");
+      assertTrue(status.transactions >= 1, `expect status.transactions = ${status.transactions} >= 1`);
 
       // It should fail to create a new trx:
       let respFailed = arango.POST("/_api/transaction/begin", data);
@@ -247,17 +190,9 @@ function testSuite() {
       resp = arango.PUT(`/_api/transaction/${resp.result.id}`, {});
       assertFalse(resp.error);
       assertEqual(200, resp.code);
-
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
     },
 
     testSoftShutdownWithAQLStreamingTrxAborted : function() {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
       // Create a streaming transaction:
       let data = { collections: {write: [cn]} };
 
@@ -271,9 +206,8 @@ function testSuite() {
       let res = arango.DELETE("/_admin/shutdown?soft=true");
       assertEqual("OK", res);
       status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-      assertEqual(0, status.AQLcursors);
-      assertEqual(1, status.transactions);
+      assertTrue(status.softShutdownOngoing, "expect status.softShutdownOngoing == true");
+      assertTrue(status.transactions >= 1, `expect status.transactions = ${status.transactions} >= 1`);
 
       // It should fail to create a new trx:
       let respFailed = arango.POST("/_api/transaction/begin", data);
@@ -287,17 +221,9 @@ function testSuite() {
       resp = arango.DELETE(`/_api/transaction/${resp.result.id}`);
       assertFalse(resp.error);
       assertEqual(200, resp.code);
-
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
     },
 
     testSoftShutdownWithAsyncRequest : function() {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
       // Create a streaming transaction:
       let data = { query: `RETURN SLEEP(15)` };
 
@@ -311,11 +237,11 @@ function testSuite() {
       assertEqual("OK", res);
       status = arango.GET("/_admin/shutdown");
       console.warn("status1:", status);
-      assertTrue(status.softShutdownOngoing);
-      assertEqual(0, status.AQLcursors);
-      assertEqual(1, status.transactions);
-      assertEqual(1, status.pendingJobs);
-      assertEqual(0, status.doneJobs);
+      assertTrue(status.softShutdownOngoing, "expect status.softShutdownOngoing == true");
+      assertTrue(status.transactions >= 1, `expect status.transactions = ${status.transactions} >= 1`);
+      // TODO: could the job already be done here? (yes if we for some reason slept for 15 secs)
+      assertEqual(1, status.pendingJobs, `expect status.pendingJobs = ${status.pendingJobs} == 1`);
+      assertEqual(0, status.doneJobs, `expect status.doneJobs = ${status.doneJobs} == 0`);
       assertFalse(status.allClear);
 
       // It should fail to create a new cursor:
@@ -329,27 +255,20 @@ function testSuite() {
       status = arango.GET("/_admin/shutdown");
       console.warn("status2:", status);
       assertTrue(status.softShutdownOngoing);
-      assertEqual(0, status.AQLcursors);
-      assertEqual(0, status.transactions);
-      assertEqual(0, status.pendingJobs);
-      assertEqual(1, status.doneJobs);
+      // TODO: could there be pending jobs? probably not if there was just one
+      // ongoing up when we started the soft shutdown (number of jobs should
+      // only be going down)
+      assertEqual(0, status.pendingJobs, `expect status.pendingJobs = ${status.pendingJobs} == 0`);
+      assertEqual(1, status.doneJobs, `expect status.doneJobs = ${status.doneJobs}" == 1`);
       assertFalse(status.allClear);
 
       // And collect the job result:
       resp = arango.PUT(`/_api/job/${resp.headers["x-arango-async-id"]}`, {});
       assertFalse(resp.error);
       assertEqual(201, resp.code);
-
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
     },
 
     testSoftShutdownWithQueuedLowPrio : function() {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
       // Create a streaming transaction:
       let op = `require("internal").wait(1); return 1;`;
 
@@ -364,17 +283,17 @@ function testSuite() {
       // Now use soft shutdown API to shut coordinator down:
       let status = arango.GET("/_admin/shutdown");
       console.warn("status0:", status);
-      assertFalse(status.softShutdownOngoing);
-      assertTrue(status.lowPrioOngoingRequests > 0);
-      assertTrue(status.lowPrioQueuedRequests > 0);
+      assertFalse(status.softShutdownOngoing, "expecting status.softShutdownOngoing == false");
+      assertTrue(status.lowPrioOngoingRequests > 0, `expect status.lowPrioOngoingRequests = ${status.lowPrioOngoingRequests} > 0`);
+      assertTrue(status.lowPrioQueuedRequests > 0, `expect status.lowPrioQueuedRequests = ${status.lowPrioQueuedRequests} > 0`);
 
       let res = arango.DELETE("/_admin/shutdown?soft=true");
       assertEqual("OK", res);
       status = arango.GET("/_admin/shutdown");
       console.warn("status1:", status);
       assertTrue(status.softShutdownOngoing);
-      assertTrue(status.lowPrioOngoingRequests > 0);
-      assertTrue(status.lowPrioQueuedRequests > 0);
+      assertTrue(status.lowPrioOngoingRequests > 0, `expect status.lowPrioOngoingRequests = ${status.lowPrioOngoingRequests} > 0`);
+      assertTrue(status.lowPrioQueuedRequests > 0, `expect status.lowPrioQueuedRequests = ${status.lowPrioQueuedRequests} > 0`);
       assertFalse(status.allClear);
 
       // Now until all jobs are done:
@@ -391,8 +310,8 @@ function testSuite() {
           assertTrue(false, "finishing jobs took too long");
         }
         console.warn("status2:", status);
-        assertTrue(status.softShutdownOngoing);
-        assertEqual(128, status.pendingJobs + status.doneJobs);
+        assertTrue(status.softShutdownOngoing, "expect status.softShutdownOngoing == true");
+        assertEqual(128, status.pendingJobs + status.doneJobs, `expect status.pendingJobs + status.doneJobs == 128; status.pendingJobs = ${status.pendingJobs}, status.doneJobs = ${status.doneJobs}`);
         assertFalse(status.allClear);
         wait(0.5);
       }
@@ -405,141 +324,13 @@ function testSuite() {
       status = arango.GET("/_admin/shutdown");
       console.warn("status3:", status);
       assertTrue(status.softShutdownOngoing);
-      assertEqual(0, status.pendingJobs);
-      assertEqual(0, status.doneJobs);
+      assertEqual(0, status.pendingJobs, `expect status.pendingJobs = ${status.pendingJobs} == 0`);
+      assertEqual(0, status.doneJobs, `expect status.doneJobs = ${status.doneJobs} == 0`);
       assertTrue(status.allClear);
-
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
     },
 
   };
 }
 
-function testSuitePregel() {
-  'use strict';
-  return {
-
-    /////////////////////////////////////////////////////////////////////////
-    /// @brief set up
-    /////////////////////////////////////////////////////////////////////////
-
-    setUp: function () {
-
-      var guck = graph_module._list();
-      var exists = guck.indexOf("demo") !== -1;
-      if (exists || db.demo_v) {
-        console.warn("Attention: graph_module gives:", guck, "or we have db.demo_v:", db.demo_v);
-        return;
-      }
-      var graph = graph_module._create(graphName);
-      db._create(vColl, { numberOfShards: 4 });
-      graph._addVertexCollection(vColl);
-      db._createEdgeCollection(eColl, {
-        numberOfShards: 4,
-        replicationFactor: 1,
-        shardKeys: ["vertex"],
-        distributeShardsLike: vColl
-      });
-
-      var rel = graph_module._relation(eColl, [vColl], [vColl]);
-      graph._extendEdgeDefinitions(rel);
-
-      var vertices = db[vColl];
-      var edges = db[eColl];
-
-
-      vertices.insert([{ _key: 'A', sssp: 3, pagerank: 0.027645934 },
-                       { _key: 'B', sssp: 2, pagerank: 0.3241496 },
-                       { _key: 'C', sssp: 3, pagerank: 0.289220 },
-                       { _key: 'D', sssp: 2, pagerank: 0.0329636 },
-                       { _key: 'E', sssp: 1, pagerank: 0.0682141 },
-                       { _key: 'F', sssp: 2, pagerank: 0.0329636 },
-                       { _key: 'G', sssp: -1, pagerank: 0.0136363 },
-                       { _key: 'H', sssp: -1, pagerank: 0.01363636 },
-                       { _key: 'I', sssp: -1, pagerank: 0.01363636 },
-                       { _key: 'J', sssp: -1, pagerank: 0.01363636 },
-                       { _key: 'K', sssp: 0, pagerank: 0.013636363 }]);
-
-      edges.insert([{ _from: vColl + '/B', _to: vColl + '/C', vertex: 'B' },
-                    { _from: vColl + '/C', _to: vColl + '/B', vertex: 'C' },
-                    { _from: vColl + '/D', _to: vColl + '/A', vertex: 'D' },
-                    { _from: vColl + '/D', _to: vColl + '/B', vertex: 'D' },
-                    { _from: vColl + '/E', _to: vColl + '/B', vertex: 'E' },
-                    { _from: vColl + '/E', _to: vColl + '/D', vertex: 'E' },
-                    { _from: vColl + '/E', _to: vColl + '/F', vertex: 'E' },
-                    { _from: vColl + '/F', _to: vColl + '/B', vertex: 'F' },
-                    { _from: vColl + '/F', _to: vColl + '/E', vertex: 'F' },
-                    { _from: vColl + '/G', _to: vColl + '/B', vertex: 'G' },
-                    { _from: vColl + '/G', _to: vColl + '/E', vertex: 'G' },
-                    { _from: vColl + '/H', _to: vColl + '/B', vertex: 'H' },
-                    { _from: vColl + '/H', _to: vColl + '/E', vertex: 'H' },
-                    { _from: vColl + '/I', _to: vColl + '/B', vertex: 'I' },
-                    { _from: vColl + '/I', _to: vColl + '/E', vertex: 'I' },
-                    { _from: vColl + '/J', _to: vColl + '/E', vertex: 'J' },
-                    { _from: vColl + '/K', _to: vColl + '/E', vertex: 'K' }]);
-    },
-
-    //////////////////////////////////////////////////////////////////////////
-    /// @brief tear down
-    //////////////////////////////////////////////////////////////////////////
-
-    tearDown: function () {
-      graph_module._drop(graphName, true);
-    },
-
-    testPageRank: function () {
-      let coordinators = getCtrlCoordinators();
-      assertTrue(coordinators.length > 0);
-      let coordinator = coordinators[0];
-
-      // Start a pregel run:
-      let pid = testAlgoStart("pagerank", { threshold: EPS / 1000, resultField: "result", store: true });
-
-      console.warn("Started pregel run:", pid);
-
-      // Now use soft shutdown API to shut coordinator down:
-      let res = arango.DELETE("/_admin/shutdown?soft=true");
-      console.warn("Shutdown got:", res);
-      assertEqual("OK", res);
-      let status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-      assertEqual(1, status.pregelConductors);
-
-      // It should fail to create a new pregel run:
-      let gotException = false;
-      try {
-        testAlgoStart("pagerank", { threshold: EPS / 10, resultField: "result", store: true });
-      } catch(e) {
-        console.warn("Got exception for new run, good!", JSON.stringify(e));
-        gotException = true;
-      }
-      assertTrue(gotException);
-
-      status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-
-      let startTime = time();
-      while (!testAlgoCheck(pid)) {
-        console.warn("Pregel still running...");
-        wait(0.2);
-        if (time() - startTime > 120) {
-          assertTrue(false, "Pregel did not finish in time.");
-        }
-      }
-
-      status = arango.GET("/_admin/shutdown");
-      assertTrue(status.softShutdownOngoing);
-
-      // And now it should shut down in due course...
-      coordinator.waitForInstanceShutdown(30);
-      restartInstance(coordinator);
-    },
-
-  };
-}
-
-jsunity.run(testSuitePregel);
 jsunity.run(testSuite);
 return jsunity.done();
